+++ src/service/service.js
@@ -3,6 +3,8 @@ const OsduEntitlementsService  = require('./entitlements');
 const OsduLegalService = require('./legal');
 const OsduQueryService = require('./query');
 const OsduStorageService = require('./storage');
+const OsduDatasetService = require('./dataset');
+const OsduSchemaService = require('./schema');
 
 /**
  * Housing class that provides named access to all OSDU endpoints via child services
@@ -23,6 +25,8 @@ class OsduService {
         this.LegalService = new OsduLegalService(osdu_client, data_partition);
         this.QueryService = new OsduQueryService(osdu_client, data_partition);
         this.StorageService = new OsduStorageService(osdu_client, data_partition);
+        this.DatasetService = new OsduDatasetService(osdu_client, data_partition);
+        this.SchemaService = new OsduSchemaService(osdu_client, data_partition);
     }
 }