+++ release-notes.md
@@ -6,6 +6,8 @@
 
 Changed versioning structure of classes to be service version based rather than OSDU release version based.
 
+Addition of dataset and schema services for R3 compatibility
+
 ## `1.0.3`
 
 **Release Date**: 2021-2-2